# docs/swagger.yaml
definitions:
...
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  request.Phone:
    properties:
      number:
        type: string
    type: object
...
  request.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  request.SignUpRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phoneNumber:
        $ref: '#/definitions/request.Phone'
      username:
        type: string
    type: object
...
  request.UpdateUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        $ref: '#/definitions/request.Phone'
      profile_picture:
        type: string
      username:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      message: {}
      statusCode:
        type: integer
    type: object
...
host: localhost:9000
info:
  contact:
    email: engdaworkyismaw9@gmail.com
    name: Engdawork yismaw
  description: This is a sample server for an e-commerce platform.
  title: E-commerce API
  version: "1.0"
paths:
...
  /user/{id}:
    get:
      description: Get a single user by id
      operationId: get-user-by-id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Get user
      tags:
      - user
  /user/delete/{id}:
    delete:
      description: delete user by id
      operationId: delete-user-by-id
      parameters:
      - description: UserID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Delete user
      tags:
      - user
  /user/list:
    get:
      description: Retrieves a list of users. Requires authentication with JWT token.
      operationId: list-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: List users
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      description: User login by email and password
      operationId: login-user
      parameters:
      - description: Login data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: Add a new user
      operationId: register-user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      summary: Insert user
      tags:
      - user
  /user/token:
    post:
      consumes:
      - application/json
      description: refresh the expired access token
      operationId: refresh-access-token
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/request.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Refresh token
      tags:
      - user
  /user/update/{id}:
    put:
      consumes:
      - application/json
      description: update user by id
      operationId: update-user-by-id
      parameters:
      - description: UserID
        in: path
        name: id
        required: true
        type: integer
      - description: Update user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Update user
      tags:
      - user
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"